+++ Scene/SceneGame.cpp
@@ -115,7 +115,8 @@ void SceneGame::Draw()
 	m_pPlayer->Draw();
 	m_pEnemy->Draw();
 	m_pPhysics->DebugDraw();
-	m_pUi->DrawHpBar(m_pPlayer->GetMaxHp(),m_pPlayer->GetNowHp(),m_pEnemy->GetMaxHp(), m_pEnemy->GetNowHp());
+	m_pUi->DrawHpBar(static_cast<float>(m_pPlayer->GetMaxHp()), static_cast<float>(m_pPlayer->GetNowHp()),
+		static_cast<float>(m_pEnemy->GetMaxHp()), static_cast<float>(m_pEnemy->GetNowHp()));
 	DrawString(0, 0, "SceneGame", GetColor(255, 255, 255));
 }
